cohere-mongodb-rag-fastapi/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
cohere==5.18.0
pymongo==4.14.1
datasets==4.0.0
tqdm==4.67.1
pandas==2.3.2
python-dotenv==1.1.1
